Remove most warnings

+++ Minecraft/include/input/Input.h
@@ -7,10 +7,10 @@ namespace Minecraft
 {
 	namespace Input
 	{
-		extern int mouseX;
-		extern int mouseY;
-		extern int deltaMouseX;
-		extern int deltaMouseY;
+		extern float mouseX;
+		extern float mouseY;
+		extern float deltaMouseX;
+		extern float deltaMouseY;
 		extern bool keyPressed[GLFW_KEY_LAST];
 		extern bool keyBeginPressData[GLFW_KEY_LAST];
